+++ packages/live-share-react/src/shared-hooks/internals/tree-node-constants.ts
@@ -1,0 +1,2 @@
+export const rawTNodeKey = "[[FluidTreeNode]]";
+export const proxiedChildrenKey = "[[FluidChildNodes]]";
